% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsFromSpreadsheet.R
\name{CredibilityIntervalFromExcel}
\alias{CredibilityIntervalFromExcel}
\title{Run Credibility Interval based on Excel inputs}
\usage{
CredibilityIntervalFromExcel(input.file1, input.file2)
}
\arguments{
\item{input.file1}{A .xlsx file for population1}

\item{input.file2}{A .xlsx file for population2}
}
\value{
list with all outputs (invisible)
}
\description{
Run Credibility Interval based on Excel inputs
}
