% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalEpiLEMMA-package.R
\docType{package}
\name{LocalEpiLEMMA}
\alias{LocalEpiLEMMA}
\title{Local Epidemic Modeling for Management & Action (LEMMA)}
\description{
LocalEpiLEMMA projects hospital utilization from COVID-19.
}
\author{
Joshua Schwab

Maintainer: Joshua Schwab \email{jschwab77@berkeley.edu}
}
