% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsFromSpreadsheet.R
\name{VaryOneParameter}
\alias{VaryOneParameter}
\title{Run Credibility Interval based on Excel inputs}
\usage{
VaryOneParameter(input.file, parameter.name = "")
}
\arguments{
\item{input.file}{A .xlsx file}
}
\value{
a ggplot object
}
\description{
Run Credibility Interval based on Excel inputs
}
