% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputsFromSpreadsheet.R
\name{CredibilityIntervalFromExcel}
\alias{CredibilityIntervalFromExcel}
\title{Run Credibility Interval based on Excel inputs}
\usage{
CredibilityIntervalFromExcel(input.file)
}
\arguments{
\item{input.file}{A .xlsx file}
}
\description{
Run Credibility Interval based on Excel inputs
}
