% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEMMA-package.R
\docType{package}
\name{LEMMA-package}
\alias{LEMMA-package}
\alias{LEMMA}
\title{The 'LEMMA' package. Local Epidemic Modeling for Management & Action}
\description{
LEMMA projects hospital utilization from COVID-19.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
Jann Paul Mattern and Mikala Caton. The Santa Cruz County COVID-19 Model, https://github.com/jpmattern/seir-covid19
}
\author{
Joshua Schwab
}
