% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEMMA-package.R
\docType{package}
\name{LEMMA}
\alias{LEMMA}
\title{Local Epidemic Modeling for Management & Action (LEMMA)}
\description{
LEMMA projects hospital utilization from COVID-19.
}
\author{
Joshua Schwab

Maintainer: Joshua Schwab \email{jschwab77@berkeley.edu}
}
